+++ axelor-supplychain/src/main/java/com/axelor/apps/supplychain/service/SaleOrderWorkflowServiceSupplychainImpl.java
@@ -17,6 +17,9 @@
  */
 package com.axelor.apps.supplychain.service;
 
+import com.axelor.apps.account.db.Invoice;
+import com.axelor.apps.account.db.repo.InvoiceRepository;
+import com.axelor.apps.account.service.invoice.InvoiceService;
 import com.axelor.apps.base.db.AppSupplychain;
 import com.axelor.apps.base.db.CancelReason;
 import com.axelor.apps.base.db.repo.PartnerRepository;
@@ -25,16 +28,24 @@ import com.axelor.apps.base.service.user.UserService;
 import com.axelor.apps.purchase.db.PurchaseOrder;
 import com.axelor.apps.purchase.db.repo.PurchaseOrderRepository;
 import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.sale.db.SaleOrderLine;
 import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.apps.sale.exception.BlockedSaleOrderException;
 import com.axelor.apps.sale.service.app.AppSaleService;
 import com.axelor.apps.sale.service.saleorder.SaleOrderWorkflowServiceImpl;
+import com.axelor.apps.stock.db.StockMove;
+import com.axelor.apps.stock.db.repo.StockMoveRepository;
+import com.axelor.apps.stock.service.StockMoveService;
 import com.axelor.apps.supplychain.service.app.AppSupplychainService;
 import com.axelor.exception.AxelorException;
+import com.axelor.i18n.I18n;
 import com.axelor.inject.Beans;
+import com.axelor.team.db.TeamTask;
+import com.axelor.team.db.repo.TeamTaskRepository;
 import com.google.inject.Inject;
 import com.google.inject.persist.Transactional;
 import java.lang.invoke.MethodHandles;
+import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -67,6 +78,10 @@ public class SaleOrderWorkflowServiceSupplychainImpl extends SaleOrderWorkflowSe
     this.accountingSituationSupplychainService = accountingSituationSupplychainService;
   }
 
+  @Inject StockMoveService stockMoveService;
+  @Inject TeamTaskRepository teamTaskRepo;
+  @Inject InvoiceService invoiceService;
+
   @Override
   @Transactional(rollbackOn = {AxelorException.class, RuntimeException.class})
   public void confirmSaleOrder(SaleOrder saleOrder) throws AxelorException {
@@ -93,12 +108,182 @@ public class SaleOrderWorkflowServiceSupplychainImpl extends SaleOrderWorkflowSe
   @Transactional
   public void cancelSaleOrder(
       SaleOrder saleOrder, CancelReason cancelReason, String cancelReasonStr) {
-    super.cancelSaleOrder(saleOrder, cancelReason, cancelReasonStr);
+
     try {
       accountingSituationSupplychainService.updateUsedCredit(saleOrder.getClientPartner());
+
+      // Cancel the associated customer deliveries
+      this.cancelAssociatedStockMoves(StockMoveRepository.ORIGIN_SALE_ORDER, saleOrder.getId());
+
+      // Cancel the associated purchase orders and its stock moves
+      List<PurchaseOrder> purchaseOrderList =
+          Beans.get(PurchaseOrderRepository.class)
+              .all()
+              .filter(
+                  "self.generatedSaleOrderId = ?1 and self.statusSelect != ?2",
+                  saleOrder.getId(),
+                  PurchaseOrderRepository.STATUS_CANCELED)
+              .fetch();
+
+      for (PurchaseOrder purchaseOrder : purchaseOrderList) {
+        this.cancelAssociatedStockMoves(
+            StockMoveRepository.ORIGIN_PURCHASE_ORDER, purchaseOrder.getId());
+        purchaseOrder.setStatusSelect(PurchaseOrderRepository.STATUS_CANCELED);
+      }
+
+      // Cancel the associated team tasks
+      List<SaleOrderLine> saleOrderLineList = saleOrder.getSaleOrderLineList();
+
+      for (SaleOrderLine saleOrderLine : saleOrderLineList) {
+        List<TeamTask> teamTaskList =
+            teamTaskRepo
+                .all()
+                .filter(
+                    "self.saleOrderLine = ?1 and self.status != ?2",
+                    saleOrderLine.getId(),
+                    TeamTaskRepository.STATUS_CANCELED)
+                .fetch();
+
+        for (TeamTask teamTask : teamTaskList) {
+          teamTask.setStatus(TeamTaskRepository.STATUS_CANCELED);
+        }
+      }
+
+      // Cancel the associated invoices
+      List<Invoice> invoiceList =
+          Beans.get(InvoiceRepository.class)
+              .all()
+              .filter(
+                  "self.saleOrder = ?1 and self.statusSelect != ?2",
+                  saleOrder.getId(),
+                  InvoiceRepository.STATUS_CANCELED)
+              .fetch();
+
+      for (Invoice invoice : invoiceList) {
+        invoiceService.cancel(invoice);
+      }
     } catch (Exception e) {
       e.printStackTrace();
     }
+
+    super.cancelSaleOrder(saleOrder, cancelReason, cancelReasonStr);
+  }
+
+  public void cancelAssociatedStockMoves(String originType, Long originId) throws AxelorException {
+
+    // Cancel the associated stock moves
+    List<StockMove> stockMoveList =
+        Beans.get(StockMoveRepository.class)
+            .all()
+            .filter(
+                "self.originTypeSelect = ?1 and self.originId = ?2 and self.statusSelect != ?3",
+                originType,
+                originId,
+                StockMoveRepository.STATUS_CANCELED)
+            .fetch();
+
+    for (StockMove stockMove : stockMoveList) {
+      stockMoveService.cancel(stockMove);
+    }
+  }
+
+  public String cancelWarningAssociatedElements(SaleOrder saleOrder) {
+    String message = null;
+    String initialMessage = I18n.get("The following associated elements will be cancelled") + " : ";
+
+    List<StockMove> stockMoveList =
+        Beans.get(StockMoveRepository.class)
+            .all()
+            .filter(
+                "self.originTypeSelect = ?1 and self.originId = ?2 and self.statusSelect != ?3",
+                StockMoveRepository.ORIGIN_SALE_ORDER,
+                saleOrder.getId(),
+                StockMoveRepository.STATUS_CANCELED)
+            .fetch();
+
+    if (stockMoveList != null && !stockMoveList.isEmpty()) {
+      String stockMoveMessage = I18n.get("Stock move(s)") + " - ";
+      message =
+          message == null
+              ? message = initialMessage + stockMoveMessage
+              : message + ", " + stockMoveMessage;
+
+      for (StockMove stockMove : stockMoveList) {
+        message = message + " " + stockMove.getStockMoveSeq();
+      }
+    }
+
+    // Associated purchase orders
+    List<PurchaseOrder> purchaseOrderList =
+        Beans.get(PurchaseOrderRepository.class)
+            .all()
+            .filter(
+                "self.generatedSaleOrderId = ?1 and self.statusSelect != ?2",
+                saleOrder.getId(),
+                PurchaseOrderRepository.STATUS_CANCELED)
+            .fetch();
+
+    if (purchaseOrderList != null && !purchaseOrderList.isEmpty()) {
+      String purchaseOrderMessage = I18n.get("Purchase order(s)") + " - ";
+      message =
+          message == null
+              ? message = initialMessage + purchaseOrderMessage
+              : message + ", " + purchaseOrderMessage;
+
+      for (PurchaseOrder purchaseOrder : purchaseOrderList) {
+        message = message + " " + purchaseOrder.getPurchaseOrderSeq();
+      }
+    }
+
+    // Associated team tasks
+    List<SaleOrderLine> saleOrderLineList = saleOrder.getSaleOrderLineList();
+
+    for (SaleOrderLine saleOrderLine : saleOrderLineList) {
+      List<TeamTask> teamTaskList =
+          teamTaskRepo
+              .all()
+              .filter(
+                  "self.saleOrderLine = ?1 and self.status != ?2",
+                  saleOrderLine.getId(),
+                  TeamTaskRepository.STATUS_CANCELED)
+              .fetch();
+
+      if (teamTaskList != null && !teamTaskList.isEmpty()) {
+        String teamTaskMessage = I18n.get("Project task(s)") + " - ";
+        message =
+            message == null
+                ? message = initialMessage + teamTaskMessage
+                : message + ", " + teamTaskMessage;
+
+        for (TeamTask teamTask : teamTaskList) {
+          message = message + " " + teamTask.getName();
+        }
+      }
+    }
+
+    // Associated invoices
+    List<Invoice> invoiceList =
+        Beans.get(InvoiceRepository.class)
+            .all()
+            .filter(
+                "self.saleOrder = ?1 and self.statusSelect != ?2",
+                saleOrder.getId(),
+                InvoiceRepository.STATUS_CANCELED)
+            .fetch();
+
+    if (invoiceList != null && !invoiceList.isEmpty()) {
+      String invoiceMessage = I18n.get("Invoice(s)") + " - ";
+      message =
+          message == null
+              ? message = initialMessage + invoiceMessage
+              : message + ", " + invoiceMessage;
+
+      for (Invoice invoice : invoiceList) {
+        message = message + " " + invoice.getInvoiceId();
+      }
+    }
+
+    return message;
   }
 
   @Override