+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Add Pack Feature in sale order.
 - FLEET : Manage rental cars and minor fixes
 - Studio: New features - Label with color,multiline string,grid column sequence,form width and order by properties. 
+- SALE ORDER : Cancel the confirmed order and its associated elements
 
 ## Improvements
 - Remove Pack Feature from Product.